pyproject.toml
[tool.poetry]
name = "ai-web-tester"
version = "0.1.0"
description = "AI-driven Web Automation Testing Framework"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.42.0"
pytest = "^8.1.1"
pytest-playwright = "^0.4.4"
langchain = "^0.1.13"
langchain-openai = "^0.1.1"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
black = "^24.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
